Futz with expiry timing.
There is now a gain factor applied to the history expiry time,
such that messages are kept in the backscroll longer the fewer
total backscroll messages there are. That is, when the
backscroll is full (15 messages), the top message will expire at
the time set. The 14th message will take longer to expire, the
13th message will take longer than the 14th, etc., all the way to
the last message, which will expire in approximately ~6 hours.

Hopefully this should help some context remain in the backscroll
for incoming users, without keeping things around forever, and
without being detrimental during active periods.