Make some experimental drawing perf improvements.
This makes things utilize CSS transforms (translates) instead of
background-position changes. This makes the browser able to paint
the complete filmstrip a single time (on image load) and then
just redraw it at the correct spot whenever it needs to. It
should trade GPU memory off for smoothness and CPU time.

Profiling this in chrome, it definitely reduces the number of
repaints on these layers, but its tough to say what effect its
actually having (FPS numbers look not all that different
between new and old. Will have to do some further testing (and
on less powerful machines).