Start moving to MP4 messages only.
Avoided this before because browsers (namely iOS safari) handled it
badly, but that's no longer the case, and videos should always work fine
provided we mark them as muted. This change also moves to
IntersectionObserver instead of Waypoints, since this should be
supported in all the browsers we care about and is a nicer API that
doesn't require a library.

GIF conversion is broken as of this commit, I have some code somewhere
(probably even on NPM?) to convert videos to GIFs, just need to hook
that back up.

This commit also removes the legacy iOS support, as that's no longer
relevant (and the previous MP4 support was being used as shorthand for
"legacy iOS user", which was a problem for what I wanted to do). After
seeing these videos on the browser, I also realized our ffmpeg settings
were really non-ideal, so I fixed those up (and things should look
better as a result).