Remove usage of ImageMagick, support for built-in HTTPS.
The ImageMagick output was unused at this point (no client should be
requesting it) so we no longer need it. The production server hasn't
used the HTTPS support in years, and it's better handled by not-JS
probably, so I removed it and updated the README with some info about
what to do instead. This simplifies the server setup code quite a bit.

Some significant restructuring of the overall protocol/usage of the
server will be coming soon, so I've avoided updating those parts of the
readme for the moment.