Start using lit-element for messages, icons, and menus.
The intention is to use it for pretty much everything, but these were
the pieces I could replace in a single part without breaking much (I
originally just wanted to replace the icon font usage, but it required a
bit more to  get working).

lit's `repeat` directive *should* handle most of the recycling behavior
we used to have. Styles may slightly differ for a few things here, as I
had to re-write some stuff to work with shadow DOM, but I tried to
preserve what actually mattered.