Make posting use HTTP instead of websockets.
Posting a bunch of data over websockets is kind of A Bad Idea as it
tends to take a long time for the browser to realize a websocket
connection is dead and there's only one socket (whereas it could
theoretically use multiple sockets to make other HTTP requests at the
same time).

The websocket endpoint is still there as of this commit, as I don't have
great handling for the protocol version changing yet, and I don't want
to break older clients. I'll remove the endpoint at some point in the
future.

I'd also like to avoid putting the resulting videos into the websockets
(and instead just hand clients an address they can use as the src), but
that change will be done separately.